Remove extra stuff from normalized HTML
And analyze the text of the normalized HTML instead of the text of the raw HTML.

This is a first, very rough cut at addressing some of the issues in #7. Since the issues pertain to the links analysis as well as the textual analysis, I just made removing the excess content a part of the standard normalization instead of a separate fallback for getting the readable body area of the page. We'll see how it works.