Reduce calls to `soup.select()`
`soup.select()` is not the most performant operation, and each call generally requires a full pass through all the nodes in the document. To reduce the cost of it (only slightly), group more selectors together so they can be run at once. (Note that compiled selectors are cached, so we don't incur a lot of extra overhead compiling -- it's the actual evaluation against the document that is costly.)