POST HTML to readability instead of proxying
The readability server can now accept POST content to parse instead of just being passed a URL to retrieve. This way, we hit S3 less and reduce a major point of network latency and redundant requests. It's actually a pretty decent performance boost, and lowers our S3 bill a bit to boot.