Increase priority for longer, contiguous changes
Look for the longest single change on the page and score highly if it's above a certain threshold (200 characters). This maxes out at an upper bound length as well (600 characters). The priority increases for each character between the bounds. The main idea here is to highlight changes that involve any significant amount of text instead of just a few words.