Update CLI interface
This is a pretty big and ugly update to `generate_task_sheets.py`, but not a real change in practice. We've been running this code in production for a while, and I kept meaning to clean it up and commit it. Instead it got messier as it accumulated changes and hacks. I never should have waited to commit -- it's time to get this into version control, and *then* clean it up.