Port old readability server from web-monitoring-changed-terms-analysis
In order to actually get automation running here, this needs to incorporate the readability server we use. I've updated the code *slightly* to account for stuff that's now built-in for more modern versions of Node.js and to use NPM instead of Yarn. It also uses the `@mozilla/readability` NPM package, which seems to be more reliably and actively updated these days (we used to install straight from git because the repo was years more up to date than the package).

Doing this has reminded me of all the things that don't work well in Mozilla's Readability. Really makes me want to take another look at Postlight or at our own stuff, which only ever got halfway done.