Bring old readability-server into this repo (#21)
Even though there are long-standing issues with Readability (see #9), we never really completed and validated a viable alternative of our own. We really need to get this tool automated, so for now that means we need to pull in Readability so this repo contains everything needed to actually run the analysis scripts.

This imports the old readability-server from the web-monitoring-changed-terms-analysis repo. I also made a few minor updates as part of doing so:
- Use the `@mozilla/readability` NPM package. We used to use their git repo directly because the NPM module seemed unmaintained (the repo had years-newer updates!). This no longer seems to be an issue.
- Use a newer version of Node.js, which has built-ins for a few things we needed 3rd-party dependencies for before.
- Update syntax to ESM.