Normalize and clean up accents
We should always have been doing this, but the problem is a lot more obvious when it comes to Spanish text. We at least need to choose a specific unicode normalization for comparisons (I've gone with NFKC, but NFKD could also work; I think it needs to be NFK-something, though). Additionally, I've gone ahead and simply removed accents altogether here, which I think is probably good for comparisons (obviously bad if we were ever going to display the resulting text).