Add some Spanish key terms (and handle accents better) (#23)
I realized yesterday that we are probably not surfacing some important changes to Spanish pages (now that we are tracking a handful) because we don’t have a Spanish version of any of our key terms. This adds a few (we should probably just translate them all).

Doing so revealed that we were not necessarily handling accents very well or very reliably, so I’ve also added code to:
1. Always compare a Unicode normalized form (NFKC in this case).
2. Remove accents from the text for comparison purposes. I think this is probably the right thing (much looser matching), but we could drop this part. It’s not as critical as the unicode normalization.