Add script for checking latest page status codes (#29)
This new workflow generates a simpler, monthly spreadsheet of just status codes for pages matching given tags or URL patterns. This is mainly meant for PEDP folks, and so it is configured to run for all URLs tagged "PEDP" when the schedule triggers it. But you can also run it manually to specify a more custom set of pages.