Narrow time range for analyzed versions
When choosing which versions to compare, try to narrow in on relevant changes by trimming versions with identical body hashes or the same HTTP error statuses from the start/end of the list of versions in the time range. I noticed in today's meeting that analysts are often not digging into the from/to version selectors on the comparison view to get an accurate time for changes, so this should help start people off with something a little more accurate. There are definitely lots of cases where it won't be as narrow as it could be since it relies on hashes. But at least much better than what we've had!