Consider Cloudflare challenges to be bad captures
We hit a lot of Cloudflare challenges on ehp.niehs.nih.gov last month that caused problems and hit them occasionally on other sites. Like AWS WAF challenges, these are not really good/valid/legitimate captures that should be considered as changed content. This adds heuristics for them to the `maybe_bad_capture()` method so we don’t use them in changes.

For more on the Cloudflare challenges, see:
- https://github.com/edgi-govdata-archiving/web-monitoring/issues/189
- https://github.com/edgi-govdata-archiving/web-monitoring-crawler/issues/32